# openapi/openapi.gen.yaml
openapi: 3.1.0
info:
  title: Loomtale Studio API
  version: 0.1.0
  description: Internal story-to-YouTube video studio API.
servers:
  - url: /api/v1
paths:
...
  /episodes:
    get:
      operationId: listEpisodes
      summary: Cursor-paginated list of a series' episodes, with aggregated draft status (no N+1)
      tags:
        - episodes
      x-min-role: viewer
      parameters:
        - name: seriesId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        '200':
          description: page of episodes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EpisodeList'
    post:
      operationId: createEpisode
      summary: Append a new, empty episode to a series
      tags:
        - episodes
      x-min-role: editor
      parameters:
        - name: seriesId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: episode created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Episode'
...
  /imports:
    get:
      operationId: listImports
      summary: Cursor-paginated list of the active tenant's imports
      tags:
        - imports
      x-min-role: viewer
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        '200':
          description: page of imports
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportList'
    post:
      operationId: createImport
      summary: Register an already-uploaded (presigned + finalized) manuscript asset as a pending import
      tags:
        - imports
      x-min-role: editor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportCreateRequest'
      responses:
        '201':
          description: import registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Import'
        '400':
          description: asset not ready, wrong MIME, or over the 10MB text cap
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /imports/{id}:
    get:
      operationId: getImport
      summary: Import status and chapter preview
      tags:
        - imports
      x-min-role: viewer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: import
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Import'
        '404':
          description: import not found in this tenant
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /imports/{id}/preview:
    post:
      operationId: previewImport
      summary: Detect encoding and split into chapters using a regex preset, without committing episodes
      tags:
        - imports
      x-min-role: editor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportPreviewRequest'
      responses:
        '200':
          description: chapter preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Import'
        '422':
          description: decoding failed or no chapters matched the preset (reported, not guessed)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /imports/{id}/commit:
    post:
      operationId: commitImport
      summary: Turn previewed chapters into episodes, optionally translating each
      tags:
        - imports
      x-min-role: editor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportCommitRequest'
      responses:
        '200':
          description: episodes created from the selected chapters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportCommitResponse'
        '409':
          description: import is not in preview status
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  schemas:
    Problem:
      type: object
      required:
        - title
        - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
...
    TargetLanguage:
      type: string
      enum:
        - en
        - vi
...
    OutlineBeat:
      type: object
      required:
        - id
        - summary
        - targetWords
      properties:
        id:
          type: string
        summary:
          type: string
        targetWords:
          type: integer
    DraftStatus:
      type: object
      properties:
        wordCount:
          type: integer
        version:
          type: integer
    Episode:
      type: object
      required:
        - id
        - seriesId
        - idx
        - title
        - outline
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        seriesId:
          type: string
          format: uuid
        idx:
          type: integer
        title:
          type: string
        outline:
          type: array
          items:
            $ref: '#/components/schemas/OutlineBeat'
        status:
          type: string
          enum:
            - planned
            - outlined
            - drafting
            - draft
            - reviewed
        drafts:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/DraftStatus'
        durationEstimateMinutes:
          type: object
          additionalProperties:
            type: number
        createdAt:
          type: string
          format: date-time
    EpisodeList:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Episode'
        nextCursor:
          type: string
...
    ChapterPreview:
      type: object
      required:
        - index
        - title
        - wordCount
      properties:
        index:
          type: integer
        title:
          type: string
        wordCount:
          type: integer
    Import:
      type: object
      required:
        - id
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        seriesId:
          type: string
          format: uuid
        assetId:
          type: string
          format: uuid
        encoding:
          type: string
        splitPreset:
          type: string
        status:
          type: string
          enum:
            - uploaded
            - preview
            - committed
            - failed
        chapters:
          type: array
          items:
            $ref: '#/components/schemas/ChapterPreview'
        errorMsg:
          type: string
        createdAt:
          type: string
          format: date-time
    ImportList:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Import'
        nextCursor:
          type: string
    ImportCreateRequest:
      type: object
      required:
        - assetId
      properties:
        assetId:
          type: string
          format: uuid
        seriesId:
          type: string
          format: uuid
    ImportPreviewRequest:
      type: object
      properties:
        splitPreset:
          type: string
          enum:
            - zh_chapter
            - en_chapter
            - vi_chuong
            - auto
          default: auto
    ImportCommitRequest:
      type: object
      required:
        - seriesId
      properties:
        seriesId:
          type: string
          format: uuid
        chapterIndexes:
          type: array
          items:
            type: integer
          description: Empty means every previewed chapter.
        translateToLang:
          $ref: '#/components/schemas/TargetLanguage'
    ImportCommitResponse:
      type: object
      required:
        - episodeIds
      properties:
        episodeIds:
          type: array
          items:
            type: string
            format: uuid
        runId:
          type: string
          format: uuid
          description: Set only when translateToLang triggered llm.translate steps.